pyproject.toml
[project]
name = "mario-rl-agent"
version = "0.1.0"
description = "Learning ML with Mario"
authors = [{ name = "Brendon Lasley", email = "lasleybrendon@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
  "torch (>=2.9.1,<3.0.0)",
  "gymnasium (>=1.2.3,<2.0.0)",
  "gym-super-mario-bros (>=7.4.0,<8.0.0)",
  "psycopg2-binary (>=2.9.11,<3.0.0)",
  "wandb (>=0.23.1,<0.24.0)",
  "numpy (<2.0)",
  "matplotlib (>=3.10.8,<4.0.0)",
  "opencv-python (<4.11)",
  "moviepy (>=2.2.1,<3.0.0)",
  "tqdm (>=4.67.1,<5.0.0)",
  "imageio[ffmpeg] (>=2.37.2,<3.0.0)",
  "pyyaml (>=6.0.3,<7.0.0)",
  "stable-baselines3 (>=2.7.1,<3.0.0)",
  "shimmy (>=2.0)",
  "rich (>=14.2.0,<15.0.0)",
  "plotly (>=6.5.0,<7.0.0)",
  "seaborn (>=0.13.2,<0.14.0)",
  "scipy (>=1.16.3,<2.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
black = "^25.12.0"
ruff = "^0.14.10"
pre-commit = "^4.5.1"
jupyter = "^1.1.1"
